Fix use numericalAmount when serializing quantity
When checking for difference between desired and actual deployment quantity may use different suffix.
Eg. when applying 2024Mi kubernetes will convert this to 2Gi. This is not a problem in it self, but it becomes a problem when the operator sdk tries to check for equality.
The operator sdk determents if it should update by removing uneccesary data from actual, serializing both actual and desired, then comparing the result. If one says 2024Mi and the other 2Gi, they will not match. To solve this problem we override the Quantity serializer, making it always use numerical amont.